+++ lib/src/configuration/better_player_translations.dart
@@ -4,6 +4,7 @@ class BetterPlayerTranslations {
   final String generalDefaultError;
   final String generalNone;
   final String generalDefault;
+  final String generalRetry;
   final String playlistLoadingNextVideo;
   final String controlsLive;
   final String controlsNextVideoIn;
@@ -16,6 +17,7 @@ class BetterPlayerTranslations {
     this.generalDefaultError = "Video can't be played",
     this.generalNone = "None",
     this.generalDefault = "Default",
+    this.generalRetry = "Retry",
     this.playlistLoadingNextVideo = "Loading next video",
     this.controlsLive = "LIVE",
     this.controlsNextVideoIn = "Next video in",
@@ -29,6 +31,7 @@ class BetterPlayerTranslations {
         generalDefaultError: "Video nie może zostać odtworzone",
         generalNone: "Brak",
         generalDefault: "Domyślne",
+        generalRetry: "Spróbuj ponownie",
         playlistLoadingNextVideo: "Ładowanie następnego filmu",
         controlsNextVideoIn: "Następne video za",
         overflowMenuPlaybackSpeed: "Szybkość odtwarzania",
@@ -41,6 +44,7 @@ class BetterPlayerTranslations {
         generalDefaultError: "无法播放视频",
         generalNone: "没有",
         generalDefault: "默认",
+        generalRetry: "重試",
         playlistLoadingNextVideo: "正在加载下一个视频",
         controlsLive: "直播",
         controlsNextVideoIn: "下一部影片n",
@@ -54,6 +58,7 @@ class BetterPlayerTranslations {
         generalDefaultError: "वीडियो नहीं चलाया जा सकता",
         generalNone: "कोई नहीं",
         generalDefault: "चूक",
+        generalRetry: "पुनः प्रयास करें",
         playlistLoadingNextVideo: "अगला वीडियो लोड हो रहा है",
         controlsLive: "लाइव",
         controlsNextVideoIn: "में अगला वीडियो",